# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "APIResponse_LabelingJobResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LabelingJobResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[LabelingJobResponse]",
        "type": "object"
      },
      "APIResponse_NoneType_": {
        "properties": {
          "data": {
            "description": "Endpoint specific payload.",
            "title": "Data",
            "type": "null"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[NoneType]",
        "type": "object"
      },
...
      "APIResponse_list_AnnotationResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[AnnotationResponse]]",
        "type": "object"
      },
      "APIResponse_list_AnnotationReviewResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AnnotationReviewResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[AnnotationReviewResponse]]",
        "type": "object"
      },
      "APIResponse_list_AnnotationTaskResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AnnotationTaskResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[AnnotationTaskResponse]]",
        "type": "object"
      },
...
      "APIResponse_list_JobAssignmentResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobAssignmentResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[JobAssignmentResponse]]",
        "type": "object"
      },
      "APIResponse_list_LabelingJobResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LabelingJobResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[LabelingJobResponse]]",
        "type": "object"
      },
      "APIResponse_list_ReviewResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReviewResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[ReviewResponse]]",
        "type": "object"
      },
...
      "AnnotationResponse": {
        "description": "Annotation representation returned by the API.",
        "properties": {
          "ai_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Confidence"
          },
          "ai_suggested_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Label"
          },
          "annotator_id": {
            "title": "Annotator Id",
            "type": "integer"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "annotator_id",
          "label",
          "notes",
          "confidence",
          "ai_suggested_label",
          "ai_confidence",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AnnotationResponse",
        "type": "object"
      },
...
      "AnnotationReviewResponse": {
        "description": "Quality review representation returned by the API.",
        "properties": {
          "annotation_id": {
            "title": "Annotation Id",
            "type": "integer"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reviewer_id": {
            "title": "Reviewer Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "annotation_id",
          "reviewer_id",
          "decision",
          "comment",
          "created_at"
        ],
        "title": "AnnotationReviewResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobAssignmentResponse": {
        "description": "Assignment representation returned by the API.",
        "properties": {
          "assigned_at": {
            "format": "date-time",
            "title": "Assigned At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "status": {
            "default": "assigned",
            "maxLength": 30,
            "title": "Status",
            "type": "string"
          },
          "worker_id": {
            "title": "Worker Id",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "worker_id",
          "id",
          "assigned_at"
        ],
        "title": "JobAssignmentResponse",
        "type": "object"
      },
...
      "LabelingJobCreate": {
        "description": "Payload used to create a labeling job.",
        "properties": {
          "annotation_type": {
            "default": "single_label_text",
            "maxLength": 40,
            "title": "Annotation Type",
            "type": "string"
          },
          "dataset_id": {
            "exclusiveMinimum": 0.0,
            "title": "Dataset Id",
            "type": "integer"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "label_options": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Label Options",
            "type": "array"
          },
          "status": {
            "default": "open",
            "maxLength": 30,
            "title": "Status",
            "type": "string"
          },
          "title": {
            "maxLength": 150,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "dataset_id"
        ],
        "title": "LabelingJobCreate",
        "type": "object"
      },
...
      "LabelingJobUpdate": {
        "description": "Payload used to update a labeling job.",
        "properties": {
          "annotation_type": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation Type"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "label_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 30,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Options"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "LabelingJobUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the AI Dataset Labeling Marketplace: datasets, labeling jobs, annotation tasks, annotations, quality review, progress tracking, dataset export and the AI annotation assistant.",
    "title": "AI Dataset Labeling Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/annotation-reviews/job/{job_id}": {
      "get": {
        "description": "Return every review recorded for a labeling job.",
        "operationId": "get_reviews_for_job_api_annotation_reviews_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationReviewResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the reviews of a labeling job",
        "tags": [
          "Annotation Reviews"
        ]
      }
    },
...
    "/api/annotations/job/{job_id}": {
      "get": {
        "description": "Return every annotation submitted for a job.",
        "operationId": "get_annotations_for_job_api_annotations_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the annotations of a labeling job",
        "tags": [
          "Annotations"
        ]
      }
    },
...
    "/api/assignments/job/{job_id}": {
      "get": {
        "description": "Return every assignment created for a job.",
        "operationId": "get_job_assignments_api_assignments_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_JobAssignmentResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the assignments of a labeling job",
        "tags": [
          "Job Assignments"
        ]
      }
    },
...
    "/api/exports/jobs/{job_id}": {
      "get": {
        "description": "Download the approved labels of a job as CSV or JSON.",
        "operationId": "export_job_api_exports_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "pattern": "^(csv|json)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_unapproved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unapproved",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export the labeled records of a labeling job",
        "tags": [
          "Exports"
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "description": "Return every labeling job, optionally filtered by status.",
        "operationId": "get_jobs_api_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_LabelingJobResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List labeling jobs",
        "tags": [
          "Labeling Jobs"
        ]
      },
      "post": {
        "description": "Create an annotation project on top of an existing dataset.",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelingJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_LabelingJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a labeling job for a dataset",
        "tags": [
          "Labeling Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "description": "Delete a labeling job and its dependent work items.",
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a labeling job",
        "tags": [
          "Labeling Jobs"
        ]
      },
      "get": {
        "description": "Return a single labeling job by identifier.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_LabelingJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retrieve one labeling job",
        "tags": [
          "Labeling Jobs"
        ]
      },
      "put": {
        "description": "Update the title, description or status of a job.",
        "operationId": "update_job_api_jobs__job_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelingJobUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_LabelingJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a labeling job",
        "tags": [
          "Labeling Jobs"
        ]
      }
    },
...
    "/api/reviews/job/{job_id}": {
      "get": {
        "description": "Return every review recorded for a job.",
        "operationId": "get_reviews_by_job_api_reviews_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_ReviewResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the reviews of a labeling job",
        "tags": [
          "Reviews"
        ]
      }
    },
...
    "/api/tasks/job/{job_id}": {
      "get": {
        "description": "Return the workload of a specific labeling job.",
        "operationId": "get_tasks_for_job_api_tasks_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationTaskResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the tasks of a labeling job",
        "tags": [
          "Annotation Tasks"
        ]
      }
    },
...
  }
}